Change CAN driver export mechanism
Move the CAN driver out from `chip/imxrt10xx` and into `chip/`. It looks
like the driver is nearly ready to build for 11xx MCUs, so I don't think
we should hide it inside the 10xx directory.

Include the driver when building for 1020 and 1060 MCUs. Do this by
using path attributes on the module. This increases the availability of
the CAN driver. Even if it's not tested on the 1020, it should still
work, and I'd like to make sure we're not designing too strictly to the
1060.

For 10xx MCUs, allow each chip-configuration module to re-export drivers
through a reexports module. Put the CAN driver in there, when
appropriate.

Test by building documentation for a 1010 MCU. Show that the package
does not have a `can` module.

    cargo doc --package=imxrt-hal --features=imxrt-ral/imxrt1011,imxrt1010 --no-deps --open

Continue testing by building documentation for a 1060 or 1020 MCU. Show
that the package has a `can` module.

    cargo doc --package=imxrt-hal --features=imxrt-ral/imxrt1062,imxrt1060 --no-deps --open
    cargo doc --package=imxrt-hal --features=imxrt-ral/imxrt1021,imxrt1020 --no-deps --open