Remove BUFLM field when referencing CAN IMASK1
The BUFLM field name is different on 1170 MCUs. That's due to SVD
construction in the imxrt-ral project. The different field name means
that we can't immediately build this CAN driver for the 1176 MCUs. We
could patch the SVDs and update the RAL, but there's a faster fix.

Study of the IMASK1 register shows that the BUFLM field is 32 bits wide.
So instead of working with the field in the register, we can just work
with the register directly. That way, we don't care about field names.

This commit changes how we interact with IMASK1. The read is equivalent,
since the field and the register are the same width. I change the way we
set the interrupt bit high for the message buffer. I'm also removing a
duplicate constant.